Move compat diag ids code into DiagnosticIDs
